refactor: move `mapStreamable` to server-only file
React's `use` should only be used with promises passed from server
components. Promises created in a client component will result in an
infinite render loop since `use` will encounter a new promise each time.
Calling `.then` results in creating a new promise, so we needed to make
sure `mapStreamable` was only called in the server to avoid creating a
new promise on the client. Furthermore, promises passed from the server
to the client are converted to a `ReactPromise` by React, which returns
`undefined` when the `.then` method is called, instead of a new
`Promise`. Probably to avoid the pitfall of creating a new promise on
each render.