feat: `<Stream>` component
Replaces the `mapStreamable` server-only utillity in favor of the
`<Stream>` component which also handles the Suspense fallback and works
in both client and server components. `mapStreamable` didn't work in
client components because its use of `Promise.then()` would result in
creating a new promise on render, a problem with React.