feat: refactor range input to fix interactions
Changing the minimum and maximum values were triggering changes but the
intent was that those changes would have to first be applied by clicking
the apply button. Furthermore, the logic to determine how to clamp on
blur and when to disable the apply button was being applied on the
parsed result, causing issues in certain edge cases.